+++ Team03/Code03/src/spa/src/sp/extractor/abstraction/follows_t_abstraction_extractor.cpp
@@ -42,7 +42,7 @@ void FollowsTAbstractionExtractor::RetrieveFromChildren(
     const std::vector<std::shared_ptr<TNode>> &children,
     std::unordered_map<TNode, StatementEntity> &t_node_stmt_ent_umap) const {
   for (int i = 0; i < (int)children.size() - 1; i++) {
-    for (int j = i + 1; j < children.size(); j++) {
+    for (int j = i + 1; j < (int)children.size(); j++) {
       auto lhs = t_node_stmt_ent_umap.find(*children[i])->second;
       auto rhs = t_node_stmt_ent_umap.find(*children[j])->second;
       follows_t_abstractions->emplace_back(lhs, rhs);
